Move state_proxies out of vtsearch/shim/ (Theme E shim rename)
The proxy layer (medias/good_votes/... _ProxyDict/_ProxyList) is the
canonical app-tier state API, not a thin adapter, so it no longer lives
under vtsearch/shim/. The shim package keeps only the genuine Flask glue
(context resolvers, persistence hooks, CoreConfig builder, plugin
families). Updated import sites, docstrings, and architecture docs.
